+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # (MODX)EvolutionCMS.snippets.ddPregReplace changelog
 
+## Version 1.1 (2018-11-25)
+* \* Attention! PHP >= 5.4 is required.
+* \* Attention! (MODX)EvolutionCMS.libraries.ddTools >= 0.23 is required.
+* \+ Added an ability to get the input string from some document field/TV.
+* \* The “str” parameter was renamed as “inputString” (with backward compatibility).
+
 ## Version 1.0.1 (2013-09-17)
 * \* Добавлен флаг 'u' для работы с юникодом.